// Angular/src/app/model/owner.ts
export class csOwner{
    OwnerID : number;
    OwnerInitials : string ;
      OwnerSurname: string ;
      OwnerPassportNumber: string ;
      OwnerMaritalStatus: string ;
      TitleID: number ;
      NationalityID: number ;
      RankID: number ;
      UserModified: string ;
      Lastmodified: Date ;
}